requirements.txt
requests==2.28.2
beautifulsoup4==4.12.2
twilio==8.1.0
python-dotenv==1.0.0
